Fix numeric comparisons for 64b.  Bug 1385

+++ src/src/expand.c
@@ -6560,7 +6560,7 @@ else
   else
     {
     while (isspace(*endptr)) endptr++;
-    if (*endptr == 0) return (int)value;
+    if (*endptr == 0) return value;
     }
   }
